Extract HiveS3TypeConfig
